Fix signedness cast

+++ include/utl/parallel_for.h
@@ -65,7 +65,7 @@ inline errors_t parallel_for(Container const& jobs, Fun&& func,
       [&](auto const idx) {
         {
           using std::begin;
-          func(*std::next(begin(jobs), idx));
+          func(*std::next(begin(jobs), static_cast<int64_t>(idx)));
         }
       },
       err_strat);
